[Device] Device/Release target now fully tested

+++ Libraries/Framework/Playdate/Template/Source/Launcher.c
@@ -38,7 +38,7 @@ int dy = 2;
 
 static int update( void* playdate_api )
 {
-  PlaydateRoutine__render__RogueInt64( (RogueInt64)playdate_api );
+  PlaydateRoutine__render__RogueInt64( (RogueInt64)(intptr_t)playdate_api );
   Rogue_check_gc();  // executes GC if new allocation threshold has been reached
 
 	return 1;
